+++ odoo/local-src/specific_magento/models/__init__.py
@@ -1,0 +1,11 @@
+from . import magento_backend
+from . import product_attribute
+from . import product_brand
+from . import product_state
+from . import product_universe
+from . import product
+from . import supplier
+from . import partner
+from . import account_invoice
+from . import sale_order
+# from . import stock_picking   # TODO do not migrate it for the moment