+++ odoo/local-src/specific_magento/models/product/__init__.py
@@ -1,0 +1,4 @@
+from . import common
+from . import backend_adapter
+from . import importer
+from . import exporter