+++ HISTORY.rst
@@ -23,6 +23,7 @@ latest (unreleased)
 
 * DBMIG-12: Install initial modules
 * DBMIG-18: Migration of Magento Connector (OCA modules)
+* DBMIG-20: Magento specific migration
 
 **Bugfixes**